drop extra file
